Refactored images storage
