+++ bot/imagesstorage/missingmongodbclient.py
@@ -1,0 +1,4 @@
+class MissingMongoDBClient(Exception):
+    def __init__(self):
+        self.message = "No MongoDB client provided"
+        super().__init__(self.message)
